+++ src/api/consumer/lib/import.js
@@ -27,6 +27,7 @@ export default (async function importAction({
   conf,
   installNpmPackages,
   withPackageJson,
+  saveDependenciesAsComponents,
   writeBitDependencies = false
 }: {
   ids: string,
@@ -41,8 +42,23 @@ export default (async function importAction({
   conf: boolean,
   installNpmPackages: boolean,
   withPackageJson: ?boolean,
+  saveDependenciesAsComponents: ?boolean,
   writeBitDependencies: ?boolean
 }): Promise<any> {
+  if (!withPackageJson) {
+    // if package.json is not written, it's impossible to install the packages and dependencies as npm packages
+    installNpmPackages = false;
+    saveDependenciesAsComponents = true;
+  }
+  if (!installNpmPackages) {
+    // if npm packages are not installed, don't install dependencies as npm packages
+    saveDependenciesAsComponents = true;
+  }
+  if (!saveDependenciesAsComponents) {
+    // according to the two 'if' statements above, installNpmPackages and withPackageJson must be true
+    writeBitDependencies = true;
+  }
+
   async function importEnvironment(consumer: Consumer): Promise<any> {
     loader.start(BEFORE_IMPORT_ENVIRONMENT);
 
@@ -100,7 +116,8 @@ export default (async function importAction({
     force,
     dist,
     conf,
-    installNpmPackages
+    installNpmPackages,
+    saveDependenciesAsComponents
   );
   const bitIds = dependencies.map(R.path(['component', 'id']));
   const bitMap = await consumer.getBitMap();