+++ src/consumer/consumer.js
@@ -271,7 +271,8 @@ export default class Consumer {
     force?: boolean = false,
     dist?: boolean = false,
     conf?: boolean = false,
-    installNpmPackages?: boolean = true
+    installNpmPackages?: boolean = true,
+    saveDependenciesAsComponents?: boolean = false
   ): Promise<> {
     const dependenciesFromBitJson = BitIds.fromObject(this.bitJson.dependencies);
     const bitMap = await this.getBitMap();
@@ -297,7 +298,8 @@ export default class Consumer {
         componentsWithDependencies: componentsAndDependenciesBitJson,
         withPackageJson,
         withBitJson: conf,
-        dist
+        dist,
+        saveDependenciesAsComponents
       });
       if (installNpmPackages) await this.installNpmPackages(componentsAndDependenciesBitJson);
     }
@@ -353,7 +355,8 @@ export default class Consumer {
     force?: boolean = false,
     dist?: boolean = false,
     conf?: boolean = false,
-    installNpmPackages?: boolean = true
+    installNpmPackages?: boolean = true,
+    saveDependenciesAsComponents?: boolean = false
   ) {
     logger.debug(`importSpecificComponents, Ids: ${rawIds.join(', ')}`);
     // $FlowFixMe - we check if there are bitIds before we call this function
@@ -368,7 +371,8 @@ export default class Consumer {
       withPackageJson,
       withBitJson: conf,
       writeBitDependencies,
-      dist
+      dist,
+      saveDependenciesAsComponents
     });
     if (installNpmPackages) await this.installNpmPackages(componentsWithDependencies);
     return { dependencies: componentsWithDependencies };
@@ -385,7 +389,8 @@ export default class Consumer {
     force?: boolean = false,
     dist?: boolean = false,
     conf?: boolean = false,
-    installNpmPackages?: boolean = true
+    installNpmPackages?: boolean = true,
+    saveDependenciesAsComponents?: boolean = false
   ): Promise<{ dependencies: ComponentWithDependencies[], envDependencies?: Component[] }> {
     loader.start(BEFORE_IMPORT_ACTION);
     if (!rawIds || R.isEmpty(rawIds)) {
@@ -397,7 +402,8 @@ export default class Consumer {
         force,
         dist,
         conf,
-        installNpmPackages
+        installNpmPackages,
+        saveDependenciesAsComponents
       );
     }
     return this.importSpecificComponents(
@@ -409,7 +415,8 @@ export default class Consumer {
       force,
       dist,
       conf,
-      installNpmPackages
+      installNpmPackages,
+      saveDependenciesAsComponents
     );
   }
 
@@ -456,7 +463,8 @@ export default class Consumer {
     withBitJson = true,
     writeBitDependencies = false,
     createNpmLinkFiles = false,
-    dist = true
+    dist = true,
+    saveDependenciesAsComponents = false
   }: {
     componentsWithDependencies: ComponentWithDependencies[],
     writeToPath?: string,
@@ -465,13 +473,18 @@ export default class Consumer {
     withBitJson?: boolean,
     writeBitDependencies?: boolean,
     createNpmLinkFiles?: boolean,
-    dist?: boolean
+    dist?: boolean,
+    saveDependenciesAsComponents?: boolean // as opposed to npm packages
   }): Promise<Component[]> {
     const bitMap: BitMap = await this.getBitMap();
     const dependenciesIdsCache = [];
+    const remotes = await this.scope.remotes();
     const writeComponentsP = componentsWithDependencies.map((componentWithDeps: ComponentWithDependencies) => {
       const bitDir = writeToPath || this.composeComponentPath(componentWithDeps.component.id);
       componentWithDeps.component.writtenPath = bitDir;
+      // if a component.scope is listed as a remote, it doesn't go to the hub and therefore it can't import dependencies as packages
+      componentWithDeps.component.dependenciesSavedAsComponents =
+        saveDependenciesAsComponents || remotes.get(componentWithDeps.component.scope);
       // AUTHORED and IMPORTED components can't be saved with multiple versions, so we can ignore the version to
       // find the component in bit.map
       const componentMap = bitMap.getComponent(componentWithDeps.component.id.toStringWithoutVersion(), false);
@@ -500,7 +513,8 @@ export default class Consumer {
     });
     const writtenComponents = await Promise.all(writeComponentsP);
 
-    const allDependenciesP = componentsWithDependencies.map((componentWithDeps) => {
+    const allDependenciesP = componentsWithDependencies.map((componentWithDeps: ComponentWithDependencies) => {
+      if (!componentWithDeps.component.dependenciesSavedAsComponents) return Promise.resolve(null);
       const writeDependenciesP = componentWithDeps.dependencies.map((dep: Component) => {
         const dependencyId = dep.id.toString();
         const depFromBitMap = bitMap.getComponent(dependencyId, false);
@@ -542,7 +556,8 @@ export default class Consumer {
 
       return Promise.all(writeDependenciesP);
     });
-    const writtenDependencies = await Promise.all(allDependenciesP);
+    const writtenDependenciesIncludesNull = await Promise.all(allDependenciesP);
+    const writtenDependencies = writtenDependenciesIncludesNull.filter(dep => dep);
 
     if (writeToPath) {
       componentsWithDependencies.forEach((componentWithDeps) => {
@@ -560,7 +575,9 @@ export default class Consumer {
       )
     );
     await bitMap.write();
-    const allComponents = [...writtenComponents, ...R.flatten(writtenDependencies)];
+    const allComponents = writtenDependencies
+      ? [...writtenComponents, ...R.flatten(writtenDependencies)]
+      : writtenComponents;
     this.linkComponents(allComponents, bitMap);
     return allComponents;
   }
@@ -848,7 +865,10 @@ export default class Consumer {
         }
         createSymlinkOrCopy(componentId, target, linkPath);
         const bound = [{ from: componentMap.rootDir, to: relativeLinkPath }];
-        const boundDependencies = component.dependencies ? writeDependenciesLinks(component, componentMap) : [];
+        const boundDependencies =
+          component.dependencies && component.dependenciesSavedAsComponents
+            ? writeDependenciesLinks(component, componentMap)
+            : [];
         const boundMissingDependencies =
           component.missingDependencies && component.missingDependencies.missingLinks
             ? writeMissingLinks(component, componentMap)
@@ -1053,16 +1073,21 @@ export default class Consumer {
 
   async installNpmPackages(componentsWithDependencies: ComponentWithDependencies[], verbose: boolean): Promise<*> {
     const componentsWithDependenciesFlatten = R.flatten(
-      componentsWithDependencies.map(oneComponentWithDependencies => [
-        oneComponentWithDependencies.component,
-        ...oneComponentWithDependencies.dependencies
-      ])
+      componentsWithDependencies.map((oneComponentWithDependencies) => {
+        return oneComponentWithDependencies.component.dependenciesSavedAsComponents
+          ? [oneComponentWithDependencies.component, ...oneComponentWithDependencies.dependencies]
+          : [oneComponentWithDependencies.component];
+      })
     );
     loader.start(BEFORE_INSTALL_NPM_DEPENDENCIES);
     const results = await Promise.all(
       componentsWithDependenciesFlatten.map((component) => {
-        if (R.isEmpty(component.packageDependencies)) return Promise.resolve();
-        return npmClient.install(component.packageDependencies, { cwd: component.writtenPath });
+        const packagesToInstall =
+          component._bitDependenciesPackages && !component.dependenciesSavedAsComponents
+            ? Object.assign(component._bitDependenciesPackages, component.packageDependencies)
+            : component.packageDependencies;
+        if (R.isEmpty(packagesToInstall)) return Promise.resolve();
+        return npmClient.install(packagesToInstall, { cwd: component.writtenPath });
       })
     );
     loader.stop();